a little validation
