+++ posthog/api/__init__.py
@@ -19,7 +19,6 @@ from . import (
     instance_status,
     integration,
     kafka_inspector,
-    media,
     organization,
     organization_domain,
     organization_invite,
@@ -32,6 +31,7 @@ from . import (
     sharing,
     site_app,
     team,
+    uploaded_media,
     user,
 )
 
@@ -105,7 +105,8 @@ projects_router.register(
 projects_router.register(
     r"property_definitions", property_definition.PropertyDefinitionViewSet, "project_property_definitions", ["team_id"]
 )
-projects_router.register(r"media", media.MediaViewSet, "project_media", ["team_id"])
+
+projects_router.register(r"uploaded_media", uploaded_media.MediaViewSet, "project_media", ["team_id"])
 
 # General endpoints (shared across CH & PG)
 router.register(r"login", authentication.LoginViewSet)
