--- posthog/api/test/fixtures/example.csv
@@ -1,2 +1,0 @@
-some,csv,content
-that,is,in-a-file
